requirements.txt
streamlit>=1.28.0
geopandas>=0.13.0
pandas>=2.0.0
matplotlib>=3.7.0
pillow>=10.0.0
requests>=2.31.0
shapely>=2.0.0
numpy>=1.24.0
scipy>=1.9.0 